+++ kukumo-plugins/kukumo-rest/src/main/java/iti/kukumo/rest/RestSupport.java
@@ -47,6 +47,7 @@ public class RestSupport {
     private RestAssuredLogger assuredLogger = new RestAssuredLogger();
     protected final Map<String, String> requestParams = new LinkedHashMap<>();
     protected final Map<String, String> queryParams = new LinkedHashMap<>();
+    protected final Map<String, String> pathParams = new LinkedHashMap<>();
     protected final Map<String, String> headers = new LinkedHashMap<>();
     protected URL baseURL;
     protected ContentType requestContentType;
@@ -71,7 +72,8 @@ public class RestSupport {
         RequestSpecification request = RestAssured.given().with()
                 .headers(headers)
                 .params(requestParams)
-                .queryParams(queryParams);
+                .queryParams(queryParams)
+                .pathParams(pathParams);
         if (attached != null) {
             request.multiPart(new MultiPartSpecBuilder(attached.getContent())
                     .fileName(attached.getName())